Move per-type application form compositions to lib-components
